Wavedrom

+++ core/views/attempts.py
@@ -23,7 +23,7 @@ from core.models import Assessment, AssessmentAttempt, CodeQuestionAttempt, Code
     CodeQuestionSubmission, TestCaseAttempt, Language, CandidateSnapshot
 from core.tasks import update_test_case_attempt_status, force_submit_assessment, compute_assessment_attempt_score, \
     detect_faces
-from core.views.utils import get_assessment_attempt_question, check_permissions_course, user_enrolled_in_course, construct_judge0_params
+from core.views.utils import get_assessment_attempt_question, check_permissions_course, user_enrolled_in_course, construct_judge0_params, vcd2wavedrom
 
 
 @login_required()
@@ -274,7 +274,6 @@ def submit_single_test_case(request, test_case_id):
                 res = requests.post(url, json=params)
                 data = res.json()
             except requests.exceptions.ConnectionError:
-                os.remove('submission.zip')
                 error_context = {
                     "result": "error",
                     "message": "Judge0 API seems to be down.",
@@ -282,10 +281,8 @@ def submit_single_test_case(request, test_case_id):
                 return Response(error_context, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
             # return error if no token
-            print(data)
             token = data.get("token")
             if not token:
-                os.remove('submission.zip')
                 error_context = {
                     "result": "error",
                     "message": "Judge0 error.",
@@ -296,7 +293,6 @@ def submit_single_test_case(request, test_case_id):
                 "result": "success",
                 "token": token,
             }
-            os.remove('submission.zip')
             return Response(context, status=status.HTTP_200_OK)
     
     except Exception as ex:
@@ -305,6 +301,11 @@ def submit_single_test_case(request, test_case_id):
             "message": f"{ex}",
         } 
         return Response(error_context, status=status.HTTP_400_BAD_REQUEST)
+    
+    finally:
+        # delete zip file
+        if os.path.exists('submission.zip'):
+            os.remove('submission.zip')
 
 
 @api_view(["GET"])
@@ -432,14 +433,16 @@ def code_question_submission(request, code_question_attempt_id):
                 url = settings.JUDGE0_URL + "/submissions/batch?base64_encoded=false"
                 res = requests.post(url, json=params)
                 data = res.json()
-                os.remove('submission.zip')
             except ConnectionError:
                 error_context = {
                     "result": "error",
                     "message": "Judge0 API seems to be down.",
                 }
-                os.remove('submission.zip')
                 return Response(error_context, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            finally:
+                # delete zip file
+                if os.path.exists('submission.zip'):
+                    os.remove('submission.zip')
             
             # retrieve tokens from judge0 response
             tokens = [x['token'] for x in data]
@@ -638,6 +641,21 @@ def vcdrom(request):
     vcd = request.POST.get('vcd')
     if vcd:
         return render(request, 'vcdrom/vcdrom.html', {'vcd': vcd})
+    else:
+        error_context = {
+            "result": "error",
+            "message": "No vcd found.",
+        }
+        return Response(error_context, status=status.HTTP_400_BAD_REQUEST)
+    
+@api_view(["POST"])
+@renderer_classes([JSONRenderer])
+@login_required()
+@groups_allowed(UserGroup.educator, UserGroup.lab_assistant, UserGroup.student)
+def wavedrom(request):
+    vcd = request.POST.get('vcd')
+    if vcd:
+        return Response({'wavedrom': vcd2wavedrom(vcd)}, status=status.HTTP_200_OK)
     else:
         error_context = {
             "result": "error",
